+++ TombEngine/Game/Lara/lara.cpp
@@ -441,186 +441,6 @@ std::function<LaraRoutineFunction> lara_collision_routines[NUM_LARA_STATES + 1] 
 	lara_default_col,//189
 };
 
-static std::optional<LaraWeaponType> GetScrolledWeaponType(const ItemInfo& item, LaraWeaponType currentWeaponType, bool getPrev)
-{
-	static const auto SCROLL_WEAPON_TYPES = std::vector<LaraWeaponType>
-	{
-		LaraWeaponType::Pistol,
-		LaraWeaponType::Shotgun,
-		LaraWeaponType::Uzi,
-		LaraWeaponType::Revolver,
-		LaraWeaponType::GrenadeLauncher,
-		LaraWeaponType::Crossbow,
-		LaraWeaponType::HarpoonGun,
-		LaraWeaponType::HK,
-		LaraWeaponType::RocketLauncher,
-	};
-
-	auto& player = GetLaraInfo(item);
-
-	// Get vector index for current weapon type.
-	auto currentIndex = std::optional<unsigned int>(std::nullopt);
-	for (int i = 0; i < SCROLL_WEAPON_TYPES.size(); i++)
-	{
-		if (SCROLL_WEAPON_TYPES[i] == currentWeaponType)
-		{
-			currentIndex = i;
-			break;
-		}
-	}
-
-	// Invalid current weapon type; return nullopt.
-	if (!currentIndex.has_value())
-		return std::nullopt;
-
-	// Next index getter.
-	auto calculateNextIndex = [getPrev](unsigned int index)
-	{
-		return (index + (getPrev ? ((unsigned int)SCROLL_WEAPON_TYPES.size() - 1) : 1)) % (unsigned int)SCROLL_WEAPON_TYPES.size();
-	};
-
-	// Get next valid weapon type in sequence.
-	unsigned int nextIndex = calculateNextIndex(*currentIndex);
-	while (nextIndex != *currentIndex)
-	{
-		auto nextWeaponType = SCROLL_WEAPON_TYPES[nextIndex];
-		if (player.Weapons[(int)nextWeaponType].Present)
-			return nextWeaponType;
-
-		nextIndex = calculateNextIndex(nextIndex);
-	}
-
-	// No valid weapon type; return nullopt.
-	return std::nullopt;
-}
-
-static void HandlePlayerQuickActions(ItemInfo& item)
-{
-	static const auto UNAVAILABLE_FLARE_STATES = std::vector<int>
-	{
-		LS_CRAWL_FORWARD,
-		LS_CRAWL_TURN_LEFT,
-		LS_CRAWL_TURN_RIGHT,
-		LS_CRAWL_BACK,
-		LS_CRAWL_TO_HANG,
-		LS_CRAWL_TURN_180
-	};
-
-	auto& player = GetLaraInfo(item);
-
-	// Handle medipack usage.
-	if (IsClicked(In::SmallMedipack) || IsClicked(In::LargeMedipack))
-	{
-		if ((item.HitPoints > 0 && item.HitPoints < LARA_HEALTH_MAX) ||
-			player.Status.Poison)
-		{
-			bool hasUsedMedipack = false;
-
-			if (IsClicked(In::SmallMedipack) &&
-				player.Inventory.TotalSmallMedipacks != 0)
-			{
-				hasUsedMedipack = true;
-
-				item.HitPoints += LARA_HEALTH_MAX / 2;
-				if (item.HitPoints > LARA_HEALTH_MAX)
-					item.HitPoints = LARA_HEALTH_MAX;
-
-				if (player.Inventory.TotalSmallMedipacks != -1)
-					player.Inventory.TotalSmallMedipacks--;
-			}
-			else if (IsClicked(In::LargeMedipack) &&
-				player.Inventory.TotalLargeMedipacks != 0)
-			{
-				hasUsedMedipack = true;
-				item.HitPoints = LARA_HEALTH_MAX;
-
-				if (player.Inventory.TotalLargeMedipacks != -1)
-					player.Inventory.TotalLargeMedipacks--;
-			}
-
-			if (hasUsedMedipack)
-			{
-				player.Status.Poison = 0;
-				SoundEffect(SFX_TR4_MENU_MEDI, nullptr, SoundEnvironment::Always);
-				Statistics.Game.HealthUsed++;
-			}
-		}
-	}
-
-	// Handle weapon scroll.
-	if (IsClicked(In::PreviousWeapon) || IsClicked(In::NextWeapon))
-	{
-		bool getPrev = IsClicked(In::PreviousWeapon);
-		auto weaponType = GetScrolledWeaponType(item, player.Control.Weapon.GunType, getPrev);
-
-		if (weaponType.has_value())
-			player.Control.Weapon.RequestGunType = *weaponType;
-	}
-
-	// Handle weapon requests.
-	if (IsClicked(In::Weapon1) && player.Weapons[(int)LaraWeaponType::Pistol].Present)
-		player.Control.Weapon.RequestGunType = LaraWeaponType::Pistol;
-
-	if (IsClicked(In::Weapon2) && player.Weapons[(int)LaraWeaponType::Shotgun].Present)
-		player.Control.Weapon.RequestGunType = LaraWeaponType::Shotgun;
-
-	if (IsClicked(In::Weapon3) && player.Weapons[(int)LaraWeaponType::Uzi].Present)
-		player.Control.Weapon.RequestGunType = LaraWeaponType::Uzi;
-
-	if (IsClicked(In::Weapon4) && player.Weapons[(int)LaraWeaponType::Revolver].Present)
-		player.Control.Weapon.RequestGunType = LaraWeaponType::Revolver;
-
-	if (IsClicked(In::Weapon5) && player.Weapons[(int)LaraWeaponType::GrenadeLauncher].Present)
-		player.Control.Weapon.RequestGunType = LaraWeaponType::GrenadeLauncher;
-
-	if (IsClicked(In::Weapon6) && player.Weapons[(int)LaraWeaponType::Crossbow].Present)
-		player.Control.Weapon.RequestGunType = LaraWeaponType::Crossbow;
-
-	if (IsClicked(In::Weapon7) && player.Weapons[(int)LaraWeaponType::HarpoonGun].Present)
-		player.Control.Weapon.RequestGunType = LaraWeaponType::HarpoonGun;
-
-	if (IsClicked(In::Weapon8) && player.Weapons[(int)LaraWeaponType::HK].Present)
-		player.Control.Weapon.RequestGunType = LaraWeaponType::HK;
-
-	if (IsClicked(In::Weapon9) && player.Weapons[(int)LaraWeaponType::RocketLauncher].Present)
-		player.Control.Weapon.RequestGunType = LaraWeaponType::RocketLauncher;
-
-	// TODO: 10th possible weapon, probably grapple gun.
-	/*if (IsClicked(In::Weapon1) && player.Weapons[(int)LaraWeaponType::].Present)
-	player.Control.Weapon.RequestGunType = LaraWeaponType::;*/
-
-	// TODO: Could theoretically remove SwitchTarget and instead do unique behaviour handling using only Look. -- Sezz 2023.07.08
-	// HACK: Handle target switch when locked on to an entity.
-	if (player.Control.HandStatus == HandStatus::WeaponReady &&
-		player.TargetEntity != nullptr)
-	{
-		if (IsClicked(In::Look))
-		{
-			ActionMap[(int)In::SwitchTarget].Update(true);
-			//ActionMap[(int)In::Look].Clear();
-		}
-		else
-		{
-			ClearAction(In::SwitchTarget);
-		}
-	}
-	else
-	{
-		ClearAction(In::SwitchTarget);
-	}
-
-	// Handle saying no.
-	if (IsClicked(In::SayNo))
-		SayNo();
-
-	// Handle flare no.
-	if (IsClicked(In::Flare))
-	{
-		if (TestState(item.Animation.ActiveState, UNAVAILABLE_FLARE_STATES))
-			SayNo();
-	}
-}
-
 void LaraControl(ItemInfo* item, CollisionInfo* coll)
 {
 	auto* lara = GetLaraInfo(item);
